pyproject.toml
[project]
name = "world-info"
version = "0.1.0"
description = "Minecraft World Repository"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]

[tool.ruff]
#Set the maximum line length to 79
#line-length=79

[tool.ruff.lint]
#Add the 'line-too-long' rule.
#extend-select = ["E501"]